+++ CHANGELOG.md
@@ -27,6 +27,7 @@ omnibus-gitlab repository.
 - Support smartcard auth against LDAP server !3006
 - Update Mattermost to 5.7.1
 - Update Prometheus alerting rules !3011
+- Allow external diffs for merge requests to be configured
 
 11.7.0